+++ src/runtime/transformers/shiki/index.ts
@@ -1,2 +1,0 @@
-export { default } from './shiki'
-export * from './highlighter'