+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 
 
+## v2.7.2
+
+[compare changes](https://github.com/nuxt/content/compare/v2.7.1...v2.7.2)
+
+### 🩹 Fixes
+
+- Remark plugins deprecated warning ([#2188](https://github.com/nuxt/content/pull/2188))
+- Allow `]` in filename code block ([#2169](https://github.com/nuxt/content/pull/2169))
+
+### 📖 Documentation
+
+- Update slots example for the `content-query` component ([#2190](https://github.com/nuxt/content/pull/2190))
+
+### ❤️  Contributors
+
+- Estéban <e.soubiran25@gmail.com>
+- Farnabaz <farnabaz@gmail.com>
+
 ## v2.7.1
 
 [compare changes](https://github.com/nuxt/content/compare/v2.7.0...v2.7.1)